Extract literal CSS/JS strings into standalone files
