+++ nmdc_runtime/api/swagger_ui/assets/style.css
@@ -1,0 +1,14 @@
+.nmdc-info {
+    padding: 1em;
+    background-color: #448aff1a;
+    border: .075rem solid #448aff;
+}
+.nmdc-info-token code {
+    font-size: x-small;
+}
+.nmdc-success {
+    color: green;
+}
+.nmdc-error {
+    color: red;
+}
